Issue-#20: Refactor 'Play' subcommand (WIP).
