+++ src/main/java/com/github/dakusui/symfonion/cli/subcommands/Help.java
@@ -1,6 +1,5 @@
 package com.github.dakusui.symfonion.cli.subcommands;
 
-import com.github.dakusui.symfonion.cli.Cli;
 import com.github.dakusui.symfonion.cli.CliRecord;
 import com.github.dakusui.symfonion.cli.Subcommand;
 import org.apache.commons.cli.HelpFormatter;
