+++ src/main/java/com/github/dakusui/symfonion/cli/subcommands/PresetSubcommand.java
@@ -1,6 +1,5 @@
 package com.github.dakusui.symfonion.cli.subcommands;
 
-import com.github.dakusui.symfonion.cli.Cli;
 import com.github.dakusui.symfonion.cli.CliRecord;
 import com.github.dakusui.symfonion.cli.Subcommand;
