+++ src/test/java/com/github/dakusui/symfonion/tests/CliTest.java
@@ -1,11 +1,5 @@
 package com.github.dakusui.symfonion.tests;
 
-import static org.junit.Assert.assertEquals;
-
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-
-import com.github.dakusui.symfonion.cli.Cli;
 import com.github.dakusui.symfonion.cli.CliRecord;
 import com.github.dakusui.symfonion.cli.subcommands.PresetSubcommand;
 import com.github.dakusui.symfonion.exceptions.CliException;
@@ -14,51 +8,62 @@ import com.github.dakusui.symfonion.testutils.TestBase;
 import org.apache.commons.cli.ParseException;
 import org.junit.Test;
 
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintStream;
+import java.nio.charset.StandardCharsets;
+
+import static org.junit.Assert.assertEquals;
+
 public class CliTest extends TestBase {
+	static CliRecord createCli(String... args) throws ParseException {
+		return new CliRecord.Builder(args).build();
+	}
 	@Test
 	public void help_01() throws CliException, ParseException {
-		Cli r = new Cli("-h");
-		assertEquals(PresetSubcommand.HELP, r.getMode());
+		CliRecord r = createCli("-h");
+		assertEquals(PresetSubcommand.HELP, r.subcommand());
 	}
 
 	@Test
 	public void help_02() throws ParseException, CliException {
-		Cli r = new Cli("--help");
-		assertEquals(PresetSubcommand.HELP, r.getMode());
+		CliRecord r = createCli("--help");
+		assertEquals(PresetSubcommand.HELP, r.subcommand());
 	}
 	
 	@Test
 	public void compile_01() throws ParseException, CliException {
 		String srcFileName = "test.json";
-		Cli r = new Cli("-c", srcFileName);
-		assertEquals(PresetSubcommand.COMPILE, r.getMode());
-		assertEquals(new File(srcFileName), r.getSourceFile());
+		CliRecord r = createCli("-c", srcFileName);
+		assertEquals(PresetSubcommand.COMPILE, r.subcommand());
+		assertEquals(new File(srcFileName), r.source());
 	}
 	
 	@Test
 	public void outportOption_01() throws ParseException, CliException {
-		Cli r = new Cli("-O", "out1=test1");
-		assertEquals("test1", r.getMidiOutDefinitions().get("out1").toString());
+		CliRecord r = createCli("-O", "out1=test1");
+		assertEquals("test1", r.midiOutRegexPatterns().get("out1").toString());
 	}
 	
 	@Test
 	public void outportOption_02() throws ParseException, CliException {
-		Cli r = new Cli("-O", "out1=test1", "-O", "out2=test2");
-		assertEquals("test1", r.getMidiOutDefinitions().get("out1").toString());
-		assertEquals("test2", r.getMidiOutDefinitions().get("out2").toString());
+		CliRecord r = createCli("-O", "out1=test1", "-O", "out2=test2");
+		assertEquals("test1", r.midiOutRegexPatterns().get("out1").toString());
+		assertEquals("test2", r.midiOutRegexPatterns().get("out2").toString());
 	}
 
 	@Test
 	public void inportOption_01() throws ParseException, CliException {
-		Cli r = new Cli("-I", "in1=test1");
-		assertEquals("test1", r.getMidiInDefinitions().get("in1").toString());
+		CliRecord r = createCli("-I", "in1=test1");
+		assertEquals("test1", r.midiInRegexPatterns().get("in1").toString());
 	}
 	
 	@Test
 	public void inportOption_02() throws ParseException, CliException {
-		Cli r = new Cli("-I", "in1=test1", "-I", "in2=test2");
-		assertEquals("test1", r.getMidiInDefinitions().get("in1").toString());
-		assertEquals("test2", r.getMidiInDefinitions().get("in2").toString());
+		CliRecord r = createCli("-I", "in1=test1", "-I", "in2=test2");
+		assertEquals("test1", r.midiInRegexPatterns().get("in1").toString());
+		assertEquals("test2", r.midiInRegexPatterns().get("in2").toString());
 	}
 
 	@Test
