+++ src/test/java/com/github/dakusui/symfonion/tests/cli/subcommands/HelpTest.java
@@ -1,6 +1,5 @@
 package com.github.dakusui.symfonion.tests.cli.subcommands;
 
-import com.github.dakusui.symfonion.cli.Cli;
 import com.github.dakusui.symfonion.cli.CliRecord;
 import com.github.dakusui.symfonion.cli.subcommands.Help;
 import com.github.dakusui.testutils.TestUtils;
