+++ src/test/java/com/github/dakusui/symfonion/tests/cli/subcommands/ListDevicesTest.java
@@ -1,9 +1,7 @@
 package com.github.dakusui.symfonion.tests.cli.subcommands;
 
-import com.github.dakusui.symfonion.cli.Cli;
 import com.github.dakusui.symfonion.cli.CliRecord;
 import com.github.dakusui.symfonion.cli.subcommands.ListDevices;
-import com.github.dakusui.symfonion.testutils.CliTestBase;
 import com.github.dakusui.symfonion.testutils.TestBase;
 import org.apache.commons.cli.ParseException;
 import org.junit.Test;
